For #262: BufferStore can now provision offset bufs, for inet4

+++ src/net/packet.cpp
@@ -7,7 +7,7 @@ using namespace net;
 
 
 Packet::~Packet(){
-  debug("<Packet> DESTRUCT packet, buf@%p \n",(uint32_t)buf_);
+  debug("<Packet> DESTRUCT packet, buf@%p \n", buf_);
   release_(buf_, capacity_);
 }
